add authorization logic to products

+++ pages/api/products/[productId].ts
@@ -1,8 +1,24 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { updateProductDto } from "@/utils/validations";
 import { prisma } from "@/db";
+import { Session, unstable_getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
-const getSingleProduct = async (req: NextApiRequest, res: NextApiResponse) => {
+const checkOwnership = (
+  productId: string,
+  session: Session | null,
+  res: NextApiResponse
+) => {
+  if (productId !== session?.user._id) {
+    res.status(403).json({ error: "Not AUthorized" });
+  }
+};
+
+const getSingleProduct = async (
+  req: NextApiRequest,
+  res: NextApiResponse,
+  session: Session | null
+) => {
   try {
     const product = await prisma.product.findUnique({
       where: { id: req.query.productId as string },
@@ -17,12 +33,17 @@ const getSingleProduct = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 };
 
-const updateProduct = async (req: NextApiRequest, res: NextApiResponse) => {
+const updateProduct = async (
+  req: NextApiRequest,
+  res: NextApiResponse,
+  session: Session | null
+) => {
   const productId = req.query.productId as string;
+  checkOwnership(productId, session, res);
+
   try {
     const result = updateProductDto.safeParse(req.body);
 
-    console.log(result);
     if (!result.success) {
       const errors = result.error.errors.map((err) => err.message);
       res.status(400).json({ error: errors });
@@ -47,8 +68,14 @@ const updateProduct = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 };
 
-const deleteProduct = async (req: NextApiRequest, res: NextApiResponse) => {
-  const { productId } = req.query;
+const deleteProduct = async (
+  req: NextApiRequest,
+  res: NextApiResponse,
+  session: Session | null
+) => {
+  const productId = req.query.productId as string;
+  checkOwnership(productId, session, res);
+
   try {
     await prisma.product.delete({
       where: { id: req.query.productId as string },
@@ -62,15 +89,31 @@ const deleteProduct = async (req: NextApiRequest, res: NextApiResponse) => {
 export default async function (req: NextApiRequest, res: NextApiResponse) {
   const { method } = req;
 
+  const session: Session | null = await unstable_getServerSession(
+    req,
+    res,
+    authOptions
+  );
+
+  if (!session) {
+    res.status(401).json({ error: "Not Authenticated" });
+    return;
+  }
+
+  if (session?.user.role === "user") {
+    res.status(403).json({ error: "Not Authorize" });
+    return;
+  }
+
   switch (method) {
     case "GET":
-      await getSingleProduct(req, res);
+      await getSingleProduct(req, res, session);
       break;
     case "PATCH":
-      await updateProduct(req, res);
+      await updateProduct(req, res, session);
       break;
     case "DELETE":
-      await deleteProduct(req, res);
+      await deleteProduct(req, res, session);
       break;
     default:
       res.status(405).json({ error: "Only GET, PATCH and DELETE" });
